Added sidebar to base.html and made songlist global
I added the sidebar code to the base.html file in order to eliminate it
from every other .html file.  By making the songlists global I was able
to increase efficency for the server as the list is now only generated
at the start and when changes are made instead of once that page is
loaded.  Need to add Roadmap to song model, allow conductor to change
roadmap, fix footer, and add slug to eboard model.